backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
redis
minio
numpy
opencv-python
Pillow
ultralytics
onnx
onnxruntime
pyzbar
lxml
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
python-multipart
